move ratification code into peachjam
