+++ africanlii/templates/africanlii/member_state_detail.html
@@ -35,7 +35,7 @@
 {% endblock %}
 {% block document-table %}
   {% if ratification_countries %}
-    {% include 'africanlii/_ratification_table.html' %}
+    {% include 'peachjam/_ratification_table.html' %}
   {% else %}
     <p>{% trans 'No documents found.' %}</p>
   {% endif %}
