--- africanlii/templates/peachjam/layouts/document_detail.html
@@ -1,32 +1,0 @@
-{% extends 'peachjam/layouts/document_detail.html' %}
-{% load i18n %}
-{% block document-tabs %}
-  {{ block.super }}
-  {% if document.work.ratification %}
-    <li class="nav-item" role="presentation">
-      <button class="nav-link"
-              data-bs-toggle="tab"
-              data-bs-target="#ratifications-tab"
-              type="button"
-              role="tab"
-              aria-controls="ratifications-tab"
-              aria-selected="false">
-        {% trans "Ratifications" %}
-        <span class="badge bg-secondary">{{ document.work.ratification.n_ratified }}</span>
-      </button>
-    </li>
-  {% endif %}
-{% endblock %}
-{% block document-tab-panes %}
-  {{ block.super }}
-  {% if document.work.ratification %}
-    <div class="tab-pane fade"
-         id="ratifications-tab"
-         role="tabpanel"
-         aria-labelledby="ratifications-tab">
-      <div class="container">
-        {% include 'africanlii/_ratifications.html' with ratification=document.work.ratification %}
-      </div>
-    </div>
-  {% endif %}
-{% endblock %}
