+++ peachjam/resources.py
@@ -50,8 +50,11 @@ from peachjam.models import (
     Locality,
     MatterType,
     Outcome,
+    Ratification,
+    RatificationCountry,
     SourceFile,
     Taxonomy,
+    Work,
     citations_processor,
 )
 from peachjam.pipelines import DOC_MIMETYPES
@@ -719,3 +722,74 @@ class AttorneyResource(resources.ModelResource):
 
     class Meta:
         model = Attorney
+
+
+class RatificationField(ForeignKeyWidget):
+    def clean(self, value, row=None, *args, **kwargs):
+        if not value:
+            raise ValueError("work frbr_uri is required")
+        work = Work.objects.filter(frbr_uri=value).first()
+        if not work:
+            raise ValueError(f'work with frbr_uri "{value}" not found')
+        ratification = Ratification.objects.update_or_create(
+            work=work,
+            defaults={
+                "source_url": row.get("source_url"),
+                "last_updated": row.get("last_updated"),
+            },
+        )[0]
+        return ratification
+
+
+class CountryField(ForeignKeyWidget):
+    def clean(self, value, row=None, *args, **kwargs):
+        if not value:
+            raise ValueError("country code is required")
+        country = Country.objects.filter(iso=value.upper()).first()
+        if not country:
+            raise ValueError(f'country with iso "{value}" not found')
+        return country
+
+
+class RatificationResource(resources.ModelResource):
+    work = fields.Field(
+        column_name="work",
+        attribute="ratification",
+        widget=RatificationField(Ratification, field="work__frbr_uri"),
+    )
+    country = fields.Field(
+        attribute="country",
+        column_name="country",
+        widget=CountryField(Country, field="name"),
+    )
+    ratification_date = fields.Field(
+        attribute="ratification_date",
+        column_name="ratification_date",
+        widget=DateWidget(),
+    )
+    deposit_date = fields.Field(
+        attribute="deposit_date",
+        column_name="deposit_date",
+        widget=DateWidget(),
+    )
+    signature_date = fields.Field(
+        attribute="signature_date",
+        column_name="signature_date",
+        widget=DateWidget(),
+    )
+    source_url = fields.Field(
+        attribute="source_url", column_name="source_url", widget=CharWidget()
+    )
+    last_updated = fields.Field(
+        attribute="last_updated",
+        column_name="last_updated",
+        widget=DateWidget(),
+    )
+
+    class Meta:
+        model = RatificationCountry
+        exclude = ("id", "ratification")
+        import_id_fields = (
+            "work",
+            "country",
+        )
