+++ peachjam/templates/peachjam/layouts/document_detail.html
@@ -128,6 +128,20 @@
                 </button>
               </li>
             {% endif %}
+            {% if document.work.ratification %}
+              <li class="nav-item" role="presentation">
+                <button class="nav-link"
+                        data-bs-toggle="tab"
+                        data-bs-target="#ratifications-tab"
+                        type="button"
+                        role="tab"
+                        aria-controls="ratifications-tab"
+                        aria-selected="false">
+                  {% trans "Ratifications" %}
+                  <span class="badge bg-secondary">{{ document.work.ratification.n_ratified }}</span>
+                </button>
+              </li>
+            {% endif %}
           {% endblock %}
         </ul>
       </div>
@@ -440,6 +454,14 @@
                role="tabpanel"
                aria-labelledby="case-history-tab">
             <div class="container">{% include 'peachjam/_case_histories.html' %}</div>
+          </div>
+        {% endif %}
+        {% if document.work.ratification %}
+          <div class="tab-pane fade"
+               id="ratifications-tab"
+               role="tabpanel"
+               aria-labelledby="ratifications-tab">
+            <div class="container">{% include 'peachjam/_ratifications.html' with ratification=document.work.ratification %}</div>
           </div>
         {% endif %}
       {% endblock %}
